Add the expand-curies command.
Add a new command, `expand-curies`, to automatically expand CURIEs found
in annotations.

That command was implemented because at some point Uberon used to
perform this kind of expansion (this was done by a Perl script that no
longer works, hence the need for a new way to do it). This is no longer
required, but the command might still come in handy, so we keep it here.