Add the "extract-orcids" command.
Add a new command, `extract-orcids`, to facilitate the injection of
ORCIDIO individuals (along with their annotations) into an ontology that
references them.

This was intended to replace a very clunky procedure (involving no less
than 4 consecutive invocations of ROBOT) in Uberon, which was ultimately
replaced by a somewhat less clunky (but still required 2 invocations)
procedure based on SPARQL. Keeping it here for posterity and possible
later reuse elsewhere.