Add the `annotate` and `normalize` commands.
Add a new `annotate` command that allows to (1) remove all ontology
annotations and (2) add a new dc:source ontology annotation derived from
the ontology version IRI. Those are two of the steps that are currently
performed by SPARQL-based processing in the ODK standard import
pipeline.

Also add a new `normalize` command that allows to (1) inject
declarations for subset properties, (2) inject declarations for synonym
types, and (3) merge logically identical axioms (axioms with different
sets of annotations). (1) and (2) are currently performed with SPARQL in
ODK workflows; (3) is automatically done as a side-effect of the SPARQL
processing.

Remove the `inject-declarations`, `merge-axiom-annotations`,
`postprocess`, and `preprocess` commands, which are now consolidated in
the two above commands.