Remove --orcid-module option.
Change the way the `extract-orcids` command work: make it send the
extracted axioms as the new output ontology in the ROBOT pipeline,
instead of allowing to either (1) write it as a separate file or (2)
merge it with the current ontology.

The problem with (1) is that it does not allow to specify the IRI of the
new ontology to write (that must be done as a separate step after the
file has been written), and the problem with (2) is that it does not
really make any sense.

The new behaviour allows to do something like:

   robot exp:extract-orcids -i input.owl --orcid-file orcidio.owl \
         annotate --ontology-iri https://example.org/myont/orcidio.owl \
         convert -f ofn -o orcid-module.owl

Also remove the possibility of using a pre-merged version of ORCIDIO
(again, it does not really make sense), and add instead the possibility
of using an online version of ORCIDIO.